olisa/major_refactor_50%_done
