--- src/core/mediator/decl.js
@@ -1,16 +1,0 @@
-class Mediator {
-    notify(system, event)
-    {
-        return
-    }
-}
-
-class ConvertToolMediator extends Mediator{
-    notify(system, event){
-        
-    }
-}
-
-export {
-
-}
