+++ src/core/systems/bridge/index.js
@@ -1,0 +1,36 @@
+/**
+ * @bridge
+ * @depricated
+ * @contains
+ * {
+ *  BridgeMonitor
+ *  BridgeObserver
+ *  RefreshObserver
+ *  TransactionCardObserver
+ * }
+ * 
+ * @bridgeEffect
+ * @contains
+ * {
+ *  useBridge
+ * }
+ * 
+ * @bridgeInput
+ * @contians
+ * {
+ *  useBridgeContext
+ *  BridgeProvider
+ *  
+ * }
+ * 
+ * @transaction
+ * @contains
+ * {
+ *  useTransactionSender
+ *  useTransactionListener
+ * }
+ */
+export * from './bridge'
+export * from './bridgeEffect'
+export * from './bridgeInput'
+export * from './transaction'
