+++ src/core/systems/chainData/index.js
@@ -1,0 +1,8 @@
+/**
+ * @prices
+ * @contains
+ * {
+ *  useETH
+ * }
+ */
+export * from './prices'
