+++ src/core/systems/event/index.js
@@ -1,22 +1,16 @@
-import { EventEmitter } from 'events'
-
 /**
- * singleton
+ * @eventManager
+ * @contains
+ * {
+ *  EventManager _events 
+ * }
+ * 
+ * @notifications
+ * @contains
+ * {
+ *  useNotification
+ *  class NotificationObserver
+ * }
  */
-class EventManager {
-    constructor(){
-        if( ! EventManager.instance){
-            this.dispatch = new EventEmitter()
-            this.dispatch.setMaxListeners(25)
-            EventManager.instance = this
-        }
-
-        return EventManager.instance
-    }
-}
-
-const _events = new EventManager()
-Object.freeze(_events)
-
-export default _events
-
+export * from './eventManager'
+export * from './notifications'
