+++ src/core/tools/data_structures/Observer.js
