+++ src/core/tools/data_structures/Validate.js
