+++ src/core/tools/utilities/zero.js
@@ -1,0 +1,8 @@
+import {getContract} from './contracts'
+
+export const controller = getContract("ZeroController")
+
+export const test = {
+    TEST_KEEPER_ADDRESS : "0x12fBc372dc2f433392CC6caB29CFBcD5082EF494",
+    SIGNALING_MULTIADDR : "/dns4/lourdehaufen.dynv6.net/tcp/443/wss/p2p-webrtc-star/"
+}
